Pull out recalculateTaxes to avoid running full EDD_Checkout object
